The Form Printout would fail inexplicably, in some complex situtations.  We were hiding a lot of failures.  I have exposed the specific errors for underlying form fields in repeat sections, if for no other reason, to help us debug this problem in the in the future.  The error that lead to this, revealed that we were not re-executing asyc evaluations properly, and were incorrectly caching results for one repeat section when evaluating another repeat section.
I removed the recursion from the repeat section component, because it was hard to debug, but it wasn't related to the issue.

I do not believe that the repeat section validator is in use.  We replace it with a "required" validation in the to-formly.pipe.ts, and that is what gets picked up.  If we want a custom message for this, we should do it, but that is erroneous code and should be removed to avoid confusion later.  I'll add a ticket.