Fixes group hide (I HOPE)
So what this does, is propertly evaluate properties in groups and repeat sections. The problem was that if you were evaluating a property WITHIN a repeat or group, you couldn't see the data outside of that group/repeat. Now, this looks at the field config in the parent's parent -- IE, the whole form