Fixes to panel and repeat
In the panel component, the HTML was getting called every time to see if we needed to hide an entire repeat section. I changed this to onInit(). This might break a condition where you are showing a repeat section dynamically based on some other value, but it greatly reduces lagginess. The other thing is onPush() in the change detector ref, which helps a little bit